lambda/usermanager/requirements.txt
argh==0.26.2
asn1crypto==0.24.0
attrs==17.4.0
backcall==0.1.0
boto3==1.7.38
botocore==1.10.38
certifi==2018.4.16
cffi==1.11.5
chalice==1.3.0
chardet==3.0.4
click==6.6
cryptography==2.2.2
decorator==4.3.0
docutils==0.14
enum34==1.1.6
httpie==0.9.9
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jmespath==0.9.3
parso==0.2.1
pathtools==0.1.2
pexpect==4.6.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pycparser==2.18
Pygments==2.2.0
PyJWT==1.6.4
requests==2.19.1
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
typing==3.6.4
urllib3==1.23
watchdog==0.8.3
wcwidth==0.1.7
